ti-linux-kernel-variscite: Align package naming with TI
This aligns package naming with TI, appending -var. It also
adds -var<Variscite version> to the package version, which should be
incremented for each Variscite version of upstream versions like 6.1.46.

It also updates update-commit.py to align with the new naming conventions
and supports passing the author as an argument.

Signed-off-by: Nate Drude <nate.d@variscite.com>